Refactor create release
