+++ src/files/manifest/compactManifestIfCore.ts
@@ -8,6 +8,11 @@ import { writeManifest } from "./writeManifest.js";
 /**
  * Reads manifest and extra files in `buildDir` compacts them in the manifest
  * and writes the resulting manifest in `buildDir`
+ *
+ * Add setup-wizard file to the manifest since packages distributed on install
+ * only include their manifest and compose.
+ * TODO: Track issue for a better solution https://github.com/dappnode/DNP_DAPPMANAGER/issues/570
+ *
  * @param buildDir `build_0.1.0`
  */
 export function compactManifestIfCore(buildDir: string): void {
