+++ src/tasks/createGithubRelease/subtasks/getNextGitTag.ts
@@ -1,0 +1,7 @@
+/**
+ * Returns the next git tag based on the next version defined in the context
+ */
+export function getNextGitTag(nextVersion?: string): string {
+  if (!nextVersion) throw Error("Missing ctx.nextVersion");
+  return `v${nextVersion}`;
+}
